small fixes for semcog deployment (#319)
* improved validation diagnostics in trip_purpose and various windows-related int32/int64 conversions

* pycodestyle

* deprecations in test code

* fix bug in handling of no viable trips case in choose_trip_destination

* add trace folder to example_mtc output folder

+++ activitysim/abm/tables/shadow_pricing.py
@@ -634,7 +634,7 @@ def buffers_for_shadow_pricing(shadow_pricing_info):
     for block_key, block_shape in block_shapes.items():
 
         # buffer_size must be int (or p2.7 long), not np.int64
-        buffer_size = int(np.prod(block_shape))
+        buffer_size = int(np.prod(block_shape, dtype=np.int64))
 
         csz = buffer_size * np.dtype(dtype).itemsize
         logger.info("allocating shared buffer %s %s buffer_size %s bytes %s (%s)" %
